Extract JUnit classes from project cloned
When visiting a java class we verify if it is a JUnit class. In case the class contains the annotation @Test, it is recognized as a JUnit class, otherwise it is just a Java class. The test classes are stored in an array of String, which will be used later to make the requests to SOCORA using test-driven search.